Connect GUI components via new event system
Separate them from the core App component.
